Designed for the storage API I want
No more mocking out an entire API, it's going
to bite me in the ass and I'm going to hate it.

Built a simple memory store for using with tests
and probably the development version until I get
a better idea of what I want.

These tests might be useful for the acceptance
tests of whatever actual backend I decide to use